fix: various small things
